requirements.txt
fastapi
uvicorn[standard]
twilio
python-dotenv
websockets
google-auth
google-cloud-secret-manager
google-cloud-firestore
audioop-lts
gunicorn
python-multipart